requirements.txt
anndata
scanpy
python-igraph
pandas
tqdm
xnetwork
scikit-learn
pynndescent
